carol/ P2P: Search box component (#5689)
* refactor search box to be reusable in other components in P2P

* remove unnecessary prop
