+++ packages/p2p/src/components/buy-sell/buy-sell-header.jsx
@@ -1,12 +1,14 @@
 import * as React from 'react';
 import PropTypes from 'prop-types';
+import debounce from 'lodash.debounce';
 import { ButtonToggle, Icon } from '@deriv/components';
+import { isDesktop } from '@deriv/shared';
 import { observer } from 'mobx-react-lite';
 import classNames from 'classnames';
 import { buy_sell } from 'Constants/buy-sell';
 import { localize } from 'Components/i18next';
 import ToggleContainer from 'Components/misc/toggle-container.jsx';
-import SearchBox from 'Components/buy-sell/search-box.jsx';
+import SearchBox from 'Components/search-box';
 import SortDropdown from 'Components/buy-sell/sort-dropdown.jsx';
 import { useStores } from 'Stores';
 import AnimationWrapper from 'Components/misc/animation-wrapper.jsx';
@@ -27,6 +29,27 @@ const BuySellHeader = ({ is_visible, table_type, setTableType }) => {
     const { buy_sell_store } = useStores();
 
     const onChangeTableType = event => setTableType(event.target.value);
+
+    const returnedFunction = debounce(() => {
+        buy_sell_store.loadMoreItems({ startIndex: 0 });
+    }, 1000);
+
+    const onClear = () => {
+        buy_sell_store.setSearchTerm('');
+        buy_sell_store.setSearchResults([]);
+    };
+
+    const onSearch = search => {
+        buy_sell_store.setSearchTerm(search.trim());
+
+        if (!search.trim()) {
+            buy_sell_store.setSearchResults([]);
+            return;
+        }
+
+        buy_sell_store.setIsLoading(true);
+        returnedFunction();
+    };
 
     React.useEffect(
         () => {
@@ -60,7 +83,11 @@ const BuySellHeader = ({ is_visible, table_type, setTableType }) => {
                     </ToggleContainer>
                 </AnimationWrapper>
                 <div className='buy-sell__header-row'>
-                    <SearchBox />
+                    <SearchBox
+                        onClear={onClear}
+                        onSearch={onSearch}
+                        placeholder={isDesktop() ? localize('Search by nickname') : localize('Search')}
+                    />
                     <SortDropdown />
                     <Icon
                         className='buy-sell__header-row--filter'
