+++ packages/p2p/src/components/search-box/index.js
@@ -1,0 +1,4 @@
+import SearchBox from './search-box.jsx';
+import './search-box.scss';
+
+export default SearchBox;
