+++ packages/p2p/src/components/search-box/search-box.scss
